+++ user/mail/assign-owa-mailbox-policy.ps1
@@ -22,7 +22,8 @@
                 "SelectSimple": {
                     "Default": "OwaMailboxPolicy-Default",
                     "No signatures": "OwaMailboxPolicy-NoSignatures",
-                    "Bookings creators": "BookingsCreators"
+                    "Bookings creators": "BookingsCreators",
+                    "Get current assignment": "GetCurrent"
                 }
             },
             "CallerName": {
@@ -41,35 +42,70 @@
 param (
     [Parameter(Mandatory = $true)]
     [String] $UserName,
-    [Parameter(Mandatory = $true)]
+    [Parameter(Mandatory = $true)][ValidateSet("OwaMailboxPolicy-Default", "OwaMailboxPolicy-NoSignatures", "BookingsCreators", "GetCurrent")]
     [string] $OwaPolicyName = "OwaMailboxPolicy-Default",
     # CallerName is tracked purely for auditing purposes
     [Parameter(Mandatory = $true)]
     [string] $CallerName
 )
 
-Write-RjRbLog -Message "Caller: '$CallerName'" -Verbose
+############################################################
+#region Variables
+#
+############################################################
 
-$Version = "1.0.0"
+$Version = "1.1.0"
+
+#endregion Variables
+
+############################################################
+
+############################################################
+#region Main Logic
+#
+############################################################
+
+Write-RjRbLog -Message "Caller: '$CallerName'" -Verbose
 Write-RjRbLog -Message "Version: $Version" -Verbose
 
 Connect-RjRbExchangeOnline
 
 try {
-    $policy = Get-OwaMailboxPolicy -Identity $OwaPolicyName -ErrorAction Stop
+    $null = Get-Mailbox -Identity $UserName -ErrorAction Stop
 }
 catch {
-    "## Could not read OWA Policy '$OwaPolicyName'. Exiting."
+    "## Could not find/read '$UserName'. Exiting"
     ""
     throw $_
 }
 
+if ($OwaPolicyName -eq "GetCurrent") {
+    try {
+        $casMailbox = Get-CasMailbox -Identity $UserName -ErrorAction Stop
+    }
+    catch {
+        "## Could not read CAS mailbox settings for '$UserName'. Exiting"
+        ""
+        throw $_
+    }
+
+    $currentPolicy = $casMailbox.OwaMailboxPolicy
+    if ([string]::IsNullOrWhiteSpace($currentPolicy)) {
+        "## OWA Mailbox Policy for '$UserName' is not set."
+    }
+    else {
+        "## OWA Mailbox Policy for '$UserName' is '$currentPolicy'."
+    }
+
+    Disconnect-ExchangeOnline -Confirm:$false | Out-Null
+    return
+}
 
 try {
-    $targetMBox = Get-Mailbox -Identity $UserName -ErrorAction Stop
+    $null = Get-OwaMailboxPolicy -Identity $OwaPolicyName -ErrorAction Stop
 }
 catch {
-    "## Could not find/read '$UserName'. Exiting"
+    "## Could not read OWA Policy '$OwaPolicyName'. Exiting."
     ""
     throw $_
 }
@@ -78,4 +114,8 @@ Set-CasMailbox -OwaMailboxPolicy $OwaPolicyName -Identity $UserName -ErrorAction
 "## OWA Mailbox Policy for '$UserName' set to '$OwaPolicyName'."
 
 Disconnect-ExchangeOnline -Confirm:$false | Out-Null
+
+#endregion Main Logic
+
+############################################################